Move DataEntryBaseRepository.kt to form library
