+++ app/src/main/java/org/dhis2/usescases/searchTrackEntity/SearchRepositoy.kt
@@ -2,7 +2,7 @@ package org.dhis2.usescases.searchTrackEntity
 
 import io.reactivex.Flowable
 import java.util.ArrayList
-import org.dhis2.data.forms.dataentry.DataEntryBaseRepository
+import org.dhis2.form.data.DataEntryBaseRepository
 import org.dhis2.form.model.FieldUiModel
 import org.dhis2.form.ui.FieldViewModelFactory
 import org.hisp.dhis.android.core.D2
