+++ form/src/main/java/org/dhis2/form/data/DataEntryBaseRepository.kt
@@ -1,6 +1,5 @@
-package org.dhis2.data.forms.dataentry
+package org.dhis2.form.data
 
-import org.dhis2.form.data.DataEntryRepository
 import org.dhis2.form.model.FieldUiModel
 import org.dhis2.form.model.SectionUiModelImpl
 import org.dhis2.form.ui.FieldViewModelFactory
